Decouple and run MLP runs for comparisons
